+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - **[Feature]** Define `DefineButtonSound` tag.
 - **[Feature]** Define `DefineButtonColorTransform` tag.
 - **[Feature]** Define `DefineVideoStream` tag.
+- **[Feature]** Define `EnablePostscript` tag.
 
 # 0.8.0 (2019-07-08)